+++ cmd/sloth/commands/validate.go
@@ -23,6 +23,7 @@ type validateCommand struct {
 	slosIncludeRegex     string
 	extraLabels          map[string]string
 	sliPluginsPaths      []string
+	sloPluginsPaths      []string
 	sloPeriodWindowsPath string
 	sloPeriod            string
 }
@@ -36,6 +37,7 @@ func NewValidateCommand(app *kingpin.Application) Command {
 	cmd.Flag("fs-include", "Filter regex to include matched discovered SLO file paths, everything else will be ignored. Exclude has preference.").Short('n').StringVar(&c.slosIncludeRegex)
 	cmd.Flag("extra-labels", "Extra labels that will be added to all the generated Prometheus rules ('key=value' form, can be repeated).").Short('l').StringMapVar(&c.extraLabels)
 	cmd.Flag("sli-plugins-path", "The path to SLI plugins (can be repeated), if not set it disable plugins support.").Short('p').StringsVar(&c.sliPluginsPaths)
+	cmd.Flag("slo-plugins-path", "The path to SLO plugins a.k.a SLO generation middlewares (can be repeated).").Short('m').StringsVar(&c.sloPluginsPaths)
 	cmd.Flag("slo-period-windows-path", "The directory path to custom SLO period windows catalog (replaces default ones).").StringVar(&c.sloPeriodWindowsPath)
 	cmd.Flag("default-slo-period", "The default SLO period windows to be used for the SLOs.").Default("30d").StringVar(&c.sloPeriod)
 
@@ -81,7 +83,12 @@ func (v validateCommand) Run(ctx context.Context, config RootConfig) error {
 	}
 
 	// Load plugins.
-	pluginRepo, err := createPluginLoader(ctx, logger, v.sliPluginsPaths)
+	pluginSLIRepo, err := createSLIPluginLoader(ctx, logger, v.sliPluginsPaths)
+	if err != nil {
+		return err
+	}
+
+	pluginSLORepo, err := createSLOPluginLoader(ctx, logger, v.sloPluginsPaths)
 	if err != nil {
 		return err
 	}
@@ -106,8 +113,8 @@ func (v validateCommand) Run(ctx context.Context, config RootConfig) error {
 	}
 
 	// Create Spec loaders.
-	promYAMLLoader := storageio.NewSlothPrometheusYAMLSpecLoader(pluginRepo, sloPeriod)
-	kubeYAMLLoader := storageio.NewK8sSlothPrometheusYAMLSpecLoader(pluginRepo, sloPeriod)
+	promYAMLLoader := storageio.NewSlothPrometheusYAMLSpecLoader(pluginSLIRepo, sloPeriod)
+	kubeYAMLLoader := storageio.NewK8sSlothPrometheusYAMLSpecLoader(pluginSLIRepo, sloPeriod)
 	openSLOYAMLLoader := storageio.NewOpenSLOYAMLSpecLoader(sloPeriod)
 
 	// For every file load the data and start the validation process:
@@ -124,9 +131,10 @@ func (v validateCommand) Run(ctx context.Context, config RootConfig) error {
 		splittedSLOsData := splitYAML(slxData)
 
 		gen := generator{
-			logger:      log.Noop,
-			windowsRepo: windowsRepo,
-			extraLabels: v.extraLabels,
+			logger:        log.Noop,
+			windowsRepo:   windowsRepo,
+			extraLabels:   v.extraLabels,
+			sloPluginRepo: pluginSLORepo,
 		}
 
 		// Prepare file validation result and start validation result for every SLO in the file.