+++ internal/plugin/slo/core/noop_v1/plugin.go
@@ -13,11 +13,11 @@ const (
 )
 
 func NewPlugin(_ json.RawMessage, _ pluginslov1.AppUtils) (pluginslov1.Plugin, error) {
-	return noopPlugin{}, nil
+	return plugin{}, nil
 }
 
-type noopPlugin struct{}
+type plugin struct{}
 
-func (p noopPlugin) ProcessSLO(ctx context.Context, request *pluginslov1.Request, result *pluginslov1.Result) error {
+func (p plugin) ProcessSLO(ctx context.Context, request *pluginslov1.Request, result *pluginslov1.Result) error {
 	return nil
 }